✨Refactor styles: separate CSS modules for DDLInputSection, DMLInputSection, QueryResultBox, and PGlitePlayground for improved organization and maintainability.
