--- frontend/apps/app/components/PGlitePage/PGlitePage.module.css
@@ -1,245 +1,0 @@
-.container {
-  font-family: var(--main-font);
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 20px;
-  line-height: 1.6;
-}
-
-.title {
-  color: var(--global-foreground);
-  border-bottom: 2px solid var(--global-border);
-  padding-bottom: 10px;
-  margin-bottom: 20px;
-}
-
-.status {
-  padding: 10px;
-  margin: 10px 0;
-  border-radius: var(--border-radius-2);
-}
-
-.loading {
-  background-color: var(--warning-background);
-  border: 1px solid var(--warning-border);
-}
-
-.success {
-  background-color: var(--success-background);
-  border: 1px solid var(--success-border);
-}
-
-.error {
-  background-color: var(--error-background);
-  border: 1px solid var(--error-border);
-}
-
-.resultContainer {
-  margin-bottom: 20px;
-}
-
-.sqlCommand {
-  font-family: var(--code-font);
-  background-color: var(--global-background-secondary);
-  padding: 5px;
-  border-radius: var(--border-radius-1);
-  margin-bottom: 8px;
-}
-
-.resultPre {
-  background-color: var(--global-background-secondary);
-  padding: 15px;
-  border-radius: var(--border-radius-2);
-  overflow-x: auto;
-  border: 1px solid var(--global-border);
-  font-family: var(--code-font);
-}
-
-/* Query Result Box Styles */
-.queryResultBox {
-  margin-bottom: 10px;
-  border-radius: var(--border-radius-2);
-  border: 1px solid var(--global-border);
-  overflow: hidden;
-}
-
-.resultHeader {
-  display: flex;
-  align-items: center;
-  padding: 10px;
-  cursor: pointer;
-  user-select: none;
-}
-
-.successHeader {
-  background-color: var(--success-background);
-  border-bottom: 1px solid var(--success-border);
-}
-
-.errorHeader {
-  background-color: var(--error-background);
-  border-bottom: 1px solid var(--error-border);
-}
-
-.statusIndicator {
-  margin-right: 10px;
-  font-size: 16px;
-}
-
-.sqlPreview {
-  flex-grow: 1;
-  font-family: var(--code-font);
-  font-size: var(--font-size-2);
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.statusMessage {
-  margin: 0 10px;
-  font-weight: 500;
-}
-
-.expandIcon {
-  font-size: 12px;
-}
-
-.resultDetails {
-  padding: 10px;
-  background-color: var(--global-background-secondary);
-}
-
-.metadataContainer {
-  margin-top: 10px;
-  padding: 8px;
-  border-top: 1px solid var(--global-border);
-  font-family: var(--main-font);
-  font-size: var(--font-size-2);
-  color: var(--global-foreground-secondary);
-}
-
-/* PGlitePlayground 専用スタイル */
-.playgroundSection {
-  margin-bottom: 2rem;
-  padding: 1.5rem;
-  border-radius: var(--border-radius-2);
-  background-color: var(--global-background-secondary);
-  border: 1px solid var(--global-border);
-}
-
-.sectionHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 1rem;
-  padding-bottom: 0.5rem;
-  border-bottom: 1px solid var(--global-border);
-}
-
-.sectionTitle {
-  font-size: var(--font-size-4);
-  font-weight: 600;
-  color: #000;
-  margin: 0;
-}
-
-.description {
-  margin-bottom: 1rem;
-  color: var(--global-foreground-secondary);
-  font-size: var(--font-size-3);
-}
-
-.codeHighlight {
-  background-color: var(--global-background-tertiary);
-  padding: 0.2rem 0.4rem;
-  border-radius: var(--border-radius-1);
-  font-family: var(--code-font);
-}
-
-.sqlTextarea {
-  width: 100%;
-  font-family: var(--code-font);
-  padding: 0.75rem;
-  border-radius: var(--border-radius-2);
-  border: 1px solid var(--global-border);
-  background-color: var(--global-background);
-  color: var(--global-foreground);
-  resize: vertical;
-  margin-bottom: 0.75rem;
-}
-
-.actionButton {
-  padding: 0.5rem 1rem;
-  border-radius: var(--border-radius-2);
-  font-weight: 500;
-  cursor: pointer;
-  transition: background-color 0.2s, border-color 0.2s;
-}
-
-.primaryButton {
-  background-color: var(--primary-accent);
-  color: var(--primary-accent-foreground);
-  border: 1px solid var(--primary-accent);
-}
-
-.primaryButton:hover {
-  background-color: var(--primary-accent-hover);
-  border-color: var(--primary-accent-hover);
-}
-
-.secondaryButton {
-  background-color: var(--global-background-secondary);
-  color: #000;
-  border: 1px solid var(--global-border);
-}
-
-.secondaryButton:hover {
-  background-color: var(--global-background-tertiary);
-}
-
-.dangerButton {
-  background-color: var(--error-background);
-  color: var(--error-foreground);
-  border: 1px solid var(--error-border);
-}
-
-.dangerButton:hover {
-  background-color: var(--error-background-hover);
-  border-color: var(--error-border-hover);
-}
-
-.dmlSection {
-  margin-bottom: 1.5rem;
-  padding: 1.25rem;
-  border-radius: var(--border-radius-2);
-  border: 1px solid var(--global-border);
-  background-color: var(--global-background-tertiary);
-}
-
-.dmlHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 1rem;
-}
-
-.dmlTitle {
-  font-size: var(--font-size-3);
-  font-weight: 600;
-  color: #000;
-  margin: 0;
-}
-
-.buttonGroup {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  margin-top: 1rem;
-}
-
-.divider {
-  height: 1px;
-  background-color: var(--global-border);
-  margin: 2rem 0;
-  width: 100%;
-}
